+++ kmpotest/kmpotest/src/iosSimulatorArm64Main/kotlin/at/asitplus/test/JUnitXMLReporter.kt
@@ -1,0 +1,3 @@
+package at.asitplus.test
+
+internal actual val target: String ="iosSimulatorArm64"