+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 Starting with May 2025, the conventions plugin does not include a Kotlin version anymore. Instead, it lets any consuming
 project set it, just as with the Android Gradle Plugin. **This is a breaking change, as there is no fallback version specified.**
 
+## 20250712
+* Kotest KMP test report generator
+
 ## 20250711
 * Kotlin 2.2.0
 * KSP-based Kotest 6.0.0.M5